ast_results/hidroh_materialistic/app/src/main/java/io/github/hidroh/materialistic/data/ItemSyncJobService.java
// isComment
package io.github.hidroh.materialistic.data;

import android.annotation.TargetApi;
import android.app.job.JobParameters;
import android.app.job.JobService;
import android.os.Build;
import android.support.annotation.NonNull;
import android.support.annotation.VisibleForTesting;
import android.text.TextUtils;
import java.util.HashMap;
import java.util.Map;
import javax.inject.Inject;
import io.github.hidroh.materialistic.ActivityModule;
import io.github.hidroh.materialistic.Injectable;

@TargetApi(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
public class isClassOrIsInterface extends JobService {

    @Inject
    RestServiceFactory isVariable;

    @Inject
    ReadabilityClient isVariable;

    private final Map<String, SyncDelegate> isVariable = new HashMap<>();

    @Override
    public void isMethod() {
        super.isMethod();
        ((Injectable) isMethod()).isMethod().isMethod(new ActivityModule(this)).isMethod(this);
    }

    @Override
    public boolean isMethod(JobParameters isParameter) {
        String isVariable = isNameExpr.isMethod(isNameExpr.isMethod());
        SyncDelegate isVariable = isMethod();
        isNameExpr.isMethod(isNameExpr, isNameExpr);
        isNameExpr.isMethod(isParameter -> {
            if (isNameExpr.isMethod(isNameExpr, isNameExpr)) {
                isMethod(isNameExpr, true);
                isNameExpr.isMethod(isNameExpr);
            }
        });
        isNameExpr.isMethod(new SyncDelegate.Job(isNameExpr.isMethod()));
        return true;
    }

    @Override
    public boolean isMethod(JobParameters isParameter) {
        String isVariable = isNameExpr.isMethod(isNameExpr.isMethod());
        if (isNameExpr.isMethod(isNameExpr)) {
            isNameExpr.isMethod(isNameExpr).isMethod();
        }
        return true;
    }

    @VisibleForTesting
    @NonNull
    SyncDelegate isMethod() {
        return new SyncDelegate(this, isNameExpr, isNameExpr);
    }
}
